goldgate(m1): fix the two live-smoke bugs the test suites could not see
The first end-to-end run of the training spine (control plane + dev NATS +
real worker + real test-prairie frames) surfaced two cross-boundary bugs,
both invisible to the per-side test suites:

1. AckWait-3600 queue wedge. A worker killed while holding an un-acked
   delivery froze the whole training queue (max_ack_pending=1) for the full
   3600s AckWait - silently, because the worker logged nothing at startup or
   on lease failure. The giant AckWait bought nothing: the keepalive thread
   already extends the ack with in_progress() every tick. AckWait drops to
   300s (recovery bound), max_deliver rises to the 20 golden constant, the
   keepalive tick derives from min(lease_ttl, ack_wait)/3, and lease-CONFLICT
   NAKs are spaced a whole AckWait apart so redeliveries outlast a crashed
   worker's stale lease instead of burning MaxDeliver in minutes. Startup,
   job lifecycle, and lease failures now log.

2. Float-ttl 400. The Go lease handlers typed ttl_seconds as int; the Python
   client json.dumps()-es its float settings ("3600.0"), which encoding/json
   refuses - every lease acquire 400'd into a silent NAK loop. Both handlers
   take float64 now, and TestTrainingLeaseAcceptsPythonWireFloats pins the
   exact Python wire bytes as raw string literals.

Also: the RareSpot frame source resolves from deployment env
(ULTRA_TRAINING_FRAME_SOURCE[_DIR]) with job params overriding - the sync
dispatch is deliberately parameterless, so the data source is an install
fact, like the fleet BisQue credentials will be.

Live proof: sync dispatch -> JetStream -> consume -> lease -> running ->
materialize 4 real frames (148 objects; prairie_dog_in_burrow correctly
unsupported/ignore) -> status-result -> server-side Gate A -> succeeded.

+++ backend/controlplane/internal/httpapi/training_lifecycle.go
@@ -564,9 +564,12 @@ func (deps ServerDeps) handleAcquireTrainingJobLease(w http.ResponseWriter, r *h
 		return
 	}
 	jobID := strings.TrimSpace(chi.URLParam(r, "job_id"))
+	// ttl_seconds is float64: the Python worker serializes its float settings
+	// as "3600.0", which encoding/json refuses to unmarshal into an int (the
+	// 400 wedged the whole training queue on first live contact).
 	var request struct {
-		WorkerID   string `json:"worker_id"`
-		TTLSeconds int    `json:"ttl_seconds"`
+		WorkerID   string  `json:"worker_id"`
+		TTLSeconds float64 `json:"ttl_seconds"`
 	}
 	if !decodeJSON(w, r, &request) {
 		return
@@ -597,8 +600,8 @@ func (deps ServerDeps) handleRenewTrainingJobLease(w http.ResponseWriter, r *htt
 	}
 	jobID := strings.TrimSpace(chi.URLParam(r, "job_id"))
 	var request struct {
-		LeaseToken string `json:"lease_token"`
-		TTLSeconds int    `json:"ttl_seconds"`
+		LeaseToken string  `json:"lease_token"`
+		TTLSeconds float64 `json:"ttl_seconds"`
 	}
 	if !decodeJSON(w, r, &request) {
 		return
@@ -635,14 +638,14 @@ func (deps ServerDeps) handleReleaseTrainingJobLease(w http.ResponseWriter, r *h
 	writeJSON(w, http.StatusOK, map[string]string{"status": "released"})
 }
 
-func trainingLeaseTTL(seconds int) time.Duration {
+func trainingLeaseTTL(seconds float64) time.Duration {
 	if seconds <= 0 {
 		// Training runs are HOURS long; the default sits far above the
 		// data-agent 300/600s defaults (section 9.2) and the OS-thread
 		// keepalive renews at TTL/3 regardless.
 		return 30 * time.Minute
 	}
-	return time.Duration(seconds) * time.Second
+	return time.Duration(seconds * float64(time.Second))
 }
 
 func (deps ServerDeps) handleAppendTrainingJobEventHTTP(w http.ResponseWriter, r *http.Request) {
